Add Marvin's improved prototype and modifications(read description)
Added Marvin's prototype and modified the result functionality to choose only ONE of the paths closeest to the delay specified. This works by analyzing all the paths in the result dictionary and comparing each key (the delay off by) to find smallest path with the delay being off by the specified delay. This works by the following comparison: if (key <= selectedKey and key != 0)